feat(Toolbar): to Function Component
