+++ packages/inline-toolbar/src/components/Toolbar/index.tsx
@@ -1,5 +1,12 @@
-/* eslint-disable react/no-array-index-key */
-import React, { ComponentType, CSSProperties, FC, ReactElement } from 'react';
+import React, {
+  Fragment,
+  ComponentType,
+  CSSProperties,
+  FC,
+  useEffect,
+  useRef,
+  useState,
+} from 'react';
 import { EditorState, getVisibleSelectionRect } from 'draft-js';
 import {
   ItalicButton,
@@ -8,21 +15,15 @@ import {
   CodeButton,
   DraftJsButtonTheme,
 } from '@draft-js-plugins/buttons';
-import { InlineToolbarPluginStore } from '../../';
+import { InlineToolbarPluginStore } from '../..';
 import { InlineToolbarPluginTheme } from '../../theme';
-
-interface OverrideContentProps {
-  getEditorState: () => EditorState;
-  setEditorState: (editorState: EditorState) => void;
-  onOverrideContent: (content: ComponentType<unknown> | undefined) => void;
-}
 
 export interface ToolbarChildrenProps {
   theme: DraftJsButtonTheme;
   getEditorState: () => EditorState;
   setEditorState: (editorState: EditorState) => void;
   onOverrideContent: (
-    content: ComponentType<OverrideContentProps> | undefined
+    content: ComponentType<ToolbarChildrenProps> | undefined
   ) => void;
 }
 
@@ -35,62 +36,42 @@ interface ToolbarProps {
   theme: InlineToolbarPluginTheme;
 }
 
-export default class Toolbar extends React.Component<ToolbarProps> {
-  static defaultProps = {
-    children: (externalProps: ToolbarChildrenProps): ReactElement => (
-      // may be use React.Fragment instead of div to improve perfomance after React 16
-      <div>
-        <ItalicButton {...externalProps} />
-        <BoldButton {...externalProps} />
-        <UnderlineButton {...externalProps} />
-        <CodeButton {...externalProps} />
-      </div>
-    ),
-  };
+export const Toolbar: FC<ToolbarProps> = ({
+  store: defaultStore,
+  children,
+  isVisible: defaultIsVisible,
+  position: defaultPosition,
+  overrideContent: defaultOverrideContent,
+  theme: defaultTheme,
+}: ToolbarProps) => {
+  const [toolbarPropsState, setToolbarPropsState] = useState<ToolbarProps>({
+    store: defaultStore,
+    isVisible: defaultIsVisible,
+    position: defaultPosition,
+    theme: defaultTheme,
+    overrideContent: defaultOverrideContent,
+  });
 
-  state: ToolbarProps = {
-    isVisible: false,
-    position: undefined,
-
-    /**
-     * If this is set, the toolbar will render this instead of the children
-     * prop and will also be shown when the editor loses focus.
-     * @type {Component}
-     */
-    overrideContent: undefined,
-  } as ToolbarProps;
-  toolbar: HTMLDivElement | null = null;
-
-  UNSAFE_componentWillMount(): void {
-    this.props.store.subscribeToItem('selection', this.onSelectionChanged);
-  }
-
-  componentWillUnmount(): void {
-    this.props.store.unsubscribeFromItem('selection', this.onSelectionChanged);
-  }
-
-  /**
-   * This can be called by a child in order to render custom content instead
-   * of the children prop. It's the responsibility of the callee to call
-   * this function again with `undefined` in order to reset `overrideContent`.
-   * @param {Component} overrideContent
-   */
-  onOverrideContent = (
-    overrideContent: ComponentType<OverrideContentProps> | undefined
+  const toolbar = useRef<HTMLDivElement>(null);
+  const onOverrideContent = (
+    newOverrideContent: ComponentType<ToolbarChildrenProps> | undefined
   ): void => {
-    this.setState({ overrideContent });
+    setToolbarPropsState((prev) => ({
+      ...prev,
+      overrideContent: newOverrideContent,
+    }));
   };
 
-  onSelectionChanged = (): void => {
+  const onSelectionChanged = (): void => {
     // need to wait a tick for window.getSelection() to be accurate
     // when focusing editor with already present selection
     setTimeout(() => {
-      if (!this.toolbar) return;
+      if (!toolbar) return;
 
       // The editor root should be two levels above the node from
       // `getEditorRef`. In case this changes in the future, we
       // attempt to find the node dynamically by traversing upwards.
-      const editorRef = this.props.store.getItem('getEditorRef')!();
+      const editorRef = toolbarPropsState.store.getItem('getEditorRef')!();
       if (!editorRef) return;
 
       // This keeps backwards compatibility with React 15
@@ -120,11 +101,11 @@ export default class Toolbar extends React.Component<ToolbarProps> {
         ancestorNode = ancestorNode.parentNode as HTMLElement;
       }
 
-      const position = {
+      const newPosition = {
         top:
           editorRoot.offsetTop -
           scrollOffset -
-          this.toolbar.offsetHeight +
+          (toolbar.current?.offsetHeight || 0) +
           (selectionRect.top - editorRootRect.top) +
           extraTopOffset,
         left:
@@ -132,21 +113,23 @@ export default class Toolbar extends React.Component<ToolbarProps> {
           (selectionRect.left - editorRootRect.left) +
           selectionRect.width / 2,
       };
-      this.setState({ position });
+      setToolbarPropsState((prevState) => ({
+        ...prevState,
+        position: newPosition,
+      }));
     });
   };
 
-  getStyle(): CSSProperties {
-    const { store } = this.props;
-    const { overrideContent, position } = this.state;
+  const getStyle = (): CSSProperties => {
+    const { overrideContent, position, store } = toolbarPropsState;
     const selection = store.getItem('getEditorState')!().getSelection();
     // overrideContent could for example contain a text input, hence we always show overrideContent
     // TODO: Test readonly mode and possibly set isVisible to false if the editor is readonly
-    const isVisible =
+    const visible =
       (!selection.isCollapsed() && selection.getHasFocus()) || overrideContent;
     const style: CSSProperties = { ...position! };
 
-    if (isVisible) {
+    if (visible) {
       style.visibility = 'visible';
       style.transform = 'translate(-50%) scale(1)';
       style.transition = 'transform 0.15s cubic-bezier(.3,1.2,.2,1)';
@@ -156,32 +139,54 @@ export default class Toolbar extends React.Component<ToolbarProps> {
     }
 
     return style;
-  }
+  };
 
-  render(): ReactElement {
-    const { theme, store } = this.props;
-    const { overrideContent: OverrideContent } = this.state;
-    const childrenProps: ToolbarChildrenProps = {
-      theme: theme.buttonStyles,
-      getEditorState: store.getItem('getEditorState')!,
-      setEditorState: store.getItem('setEditorState')!,
-      onOverrideContent: this.onOverrideContent,
-    };
+  const childrenProps: ToolbarChildrenProps = {
+    theme: toolbarPropsState.theme.buttonStyles,
+    getEditorState: toolbarPropsState.store.getItem('getEditorState')!,
+    setEditorState: toolbarPropsState.store.getItem('setEditorState')!,
+    onOverrideContent,
+  };
 
-    return (
-      <div
-        className={theme.toolbarStyles.toolbar}
-        style={this.getStyle()}
-        ref={(element) => {
-          this.toolbar = element;
-        }}
-      >
-        {OverrideContent ? (
-          <OverrideContent {...childrenProps} />
-        ) : (
-          this.props.children!(childrenProps)
-        )}
-      </div>
-    );
-  }
-}
+  useEffect(() => {
+    toolbarPropsState.store.subscribeToItem('selection', onSelectionChanged);
+  }, []);
+
+  useEffect(
+    () => () => {
+      toolbarPropsState.store.unsubscribeFromItem(
+        'selection',
+        onSelectionChanged
+      );
+    },
+    []
+  );
+  const { overrideContent: OverrideContent } = toolbarPropsState;
+
+  return (
+    <div
+      className={toolbarPropsState.theme.toolbarStyles.toolbar}
+      style={getStyle()}
+      ref={toolbar}
+    >
+      {OverrideContent ? (
+        <OverrideContent {...childrenProps} />
+      ) : (
+        <>
+          {children ? (
+            children(childrenProps)
+          ) : (
+            <Fragment>
+              <ItalicButton {...childrenProps} />
+              <BoldButton {...childrenProps} />
+              <UnderlineButton {...childrenProps} />
+              <CodeButton {...childrenProps} />
+            </Fragment>
+          )}
+        </>
+      )}
+    </div>
+  );
+};
+
+export default Toolbar;
